inspect-psql/dash_app/requirements.txt
pandas
dash
psycopg2-binary
sqlalchemy
plotly